Minor changes to keep platforms consistent.
