+++ src/hotspot/cpu/aarch64/continuationHelper_aarch64.inline.hpp
@@ -55,17 +55,6 @@ static inline intptr_t** link_address(const frame& f) {
             : (intptr_t**)(f.unextended_sp() + f.cb()->frame_size() - frame::sender_sp_offset);
 }
 
-static void patch_callee_link(const frame& f, intptr_t* fp) {
-  DEBUG_ONLY(intptr_t* orig = *ContinuationHelper::Frame::callee_link_address(f));
-  *ContinuationHelper::Frame::callee_link_address(f) = fp;
-}
-
-static void patch_callee_link_relative(const frame& f, intptr_t* fp) {
-  intptr_t* la = (intptr_t*)ContinuationHelper::Frame::callee_link_address(f);
-  intptr_t new_value = fp - la;
-  *la = new_value;
-}
-
 inline int ContinuationHelper::frame_align_words(int size) {
 #ifdef _LP64
   return size & 1;
